Bugfix: 3744
The grub_file_open() function uses the global variable grub_errno
as a flag to indicate that an error occurred during the execution
of the grub_file_get_device_name() function.  That function can succeed,
but the error test fail, if grub_errno was set to an error value before
the function call.  The fix is to initialize grub_errno before the
function call.