Bugfix 5399: Don't admit members that are not "fresh" to a RAID array.
Previously, all members of an array would be added, even those that
had previously been removed from the array and which may hold
out-of-date data.  With this fix, the array assembly code consults the
"events" field of the RAID superblock of each member.  Only the member
that has, or the members that share, the highest value in their
"events" field are admitted to the array.  Members with lower value
"events" field are considered "not fresh" and are not admitted.

This is the same logic that the kernel "md" code uses to admit members
to a RAID array.